requirements.txt
torch
torchvision
torchaudio
transformers
gradio
scikit-learn
matplotlib
tqdm
pillow
numpy
